docs(examples-playwright): wait for interceptors to load
