--- examples/with-playwright/src/providers/InterceptorProvider.tsx
@@ -1,19 +1,0 @@
-'use client';
-
-import { PropsWithChildren, useEffect } from 'react';
-
-import { loadInterceptors } from '../../tests/interceptors';
-
-type Props = PropsWithChildren;
-
-function InterceptorProvider({ children }: Props) {
-  useEffect(() => {
-    if (process.env.NODE_ENV === 'development') {
-      void loadInterceptors();
-    }
-  }, []);
-
-  return children;
-}
-
-export default InterceptorProvider;
