+++ examples/with-playwright/src/services/github.ts
@@ -1,6 +1,9 @@
+import { cache } from 'react';
 import type { JSONValue } from 'zimic';
 
-const NEXT_PUBLIC_GITHUB_API_BASE_URL = process.env.NEXT_PUBLIC_GITHUB_API_BASE_URL ?? '';
+import { waitForLoadedInterceptors } from '../providers/interceptors/utils';
+
+export const GITHUB_API_BASE_URL = process.env.NEXT_PUBLIC_GITHUB_API_BASE_URL ?? '';
 
 export type GitHubRepository = JSONValue<{
   id: number;
@@ -10,9 +13,11 @@ export type GitHubRepository = JSONValue<{
   owner: { login: string };
 }>;
 
-export async function fetchGitHubRepository(ownerName: string, repositoryName: string) {
+export const fetchGitHubRepository = cache(async (ownerName: string, repositoryName: string) => {
+  await waitForLoadedInterceptors();
+
   try {
-    const repositoryURL = `${NEXT_PUBLIC_GITHUB_API_BASE_URL}/repos/${ownerName}/${repositoryName}`;
+    const repositoryURL = `${GITHUB_API_BASE_URL}/repos/${ownerName}/${repositoryName}`;
     const repositoryResponse = await fetch(repositoryURL, {
       cache: process.env.NODE_ENV === 'development' ? 'no-store' : 'default',
     });
@@ -27,4 +32,4 @@ export async function fetchGitHubRepository(ownerName: string, repositoryName: s
     console.error(error);
     return null;
   }
-}
+});
